MAINT: doc fixes and renaming
