+++ enigma-principal/app/src/boot_network/mod.rs
@@ -1,2 +1,2 @@
-pub mod keys_provider_http;
+pub mod km_http_server;
 pub mod principal_manager;
